Change storage of colors and timestamps to uint
Surfel color and timestamps are now stored jointly in a uint4 vector.
This commit involves changing of related texture and VBO formats.

+++ Core/src/Shaders/draw_feedback.vert
@@ -19,7 +19,7 @@
 #version 430 core
 
 layout (location = 0) in vec4 position;
-layout (location = 1) in vec4 color;
+layout (location = 1) in uvec4 color;
 layout (location = 2) in vec4 normal;
 
 uniform mat4 MVP;
